requirement.txt
pvporcupine
pyaudio
numpy
pyttsx
soundfile
sounddevice
vosk
pycaw
comtypes
rapidfuzz
psutil
requests
python-speech-features 
scikit-learn
joblib
dotenv
faster_whisper
tkinter
ttkbootstrap
whispercpp
pipwin
py-cpuinfo
python-dotenv 
pvleopard
noicereduce
